carwash_api/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-django==4.7.0